Backend/requirements.txt
fastapi[all]
motor
pydantic
uvicorn
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client 
python-multipart
google-generativeai 
python-dotenv 
python-docx
#laavanjan's libraries
langchain==0.3.25
langchain_community==0.3.25
langchain_core==0.3.65
langchain_groq==0.3.2
langchain_huggingface==0.3.0
pymongo
python-dotenv
Requests
faiss-cpu
sentence-transformers
pypdf
python-jose[cryptography]
PyMuPDF==1.23.8
jspdf
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
pickle-mixin>=1.0.2